quic: fix compiler error
Fixes: https://github.com/nodejs/quic/issues/89
PR-URL: https://github.com/nodejs/quic/pull/92

+++ src/node_quic_stream.cc
@@ -455,7 +455,9 @@ void QuicStreamShutdown(const FunctionCallbackInfo<Value>& args) {
   uint64_t code = ExtractErrorCode(env, args[0]);
   USE(args[1]->Uint32Value(env->context()).To(&family));
 
-  stream->Shutdown(family == QUIC_ERROR_APPLICATION ? code : NGTCP2_NO_ERROR);
+  stream->Shutdown(
+      family == QUIC_ERROR_APPLICATION ?
+          code : static_cast<uint64_t>(NGTCP2_NO_ERROR));
 }
 }  // namespace
